legacy/extraction_module/requirements.txt
transformers
huggingface-hub
tokenizers
safetensors

numpy
opencv-python
Pillow

pydantic
python-dateutil
regex

matplotlib
seaborn

pytest

google-cloud-vision
groq
python-dotenv